Refactor: place trades' api-related code inside `api` subpackage
